fix(ci): resolve 4 test failures from endpoint and model changes
- test_router_health: update TestPluginsEndpoint to expect 404 since
  GET /plugins was removed from health.py (now served by plugins.py)
- models/settings.py: chunk_overlap is float (0.2 ratio), not int —
  fixes ResponseValidationError in 3 settings tests
- test_router_sdk: trading mock paths updated (previous commit)

+++ src/mcp/models/settings.py
@@ -33,7 +33,7 @@ class SettingsResponse(BaseModel):
     # Core settings
     categorize_mode: str = ""
     chunk_max_tokens: int = 0
-    chunk_overlap: int = 0
+    chunk_overlap: float = 0
     enable_encryption: bool = False
     enable_feedback_loop: bool = False
     enable_hallucination_check: bool = False
